fixed error with float32 (TypeError: Object of type float32 is not JSON serializable), because downcast='float' converts Decimal to float32

+++ project/bookkeeping/tests/test_lib_make_dataframe.py
@@ -70,7 +70,7 @@ def test_month_dtypes(data, types):
         ])
     actual = make_dataframe.MakeDataFrame(year=1999, data=data, types=types).data
 
-    assert actual.dtypes['T1'] == np.float32
+    assert actual.dtypes['T1'] == np.float64
 
 
 @pytest.mark.parametrize('data', [([]), (None)])
